# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sports2 API
  description: |
    REST API for the Sports2 collegiate baseball scouting and team management platform.

    ## Authentication
    All endpoints (except public ones) require:
    - `Authorization: Bearer <JWT>` header
    - `x-csrf-token` header on state-changing requests (POST, PUT, PATCH, DELETE)

    Get a CSRF token first: `GET /api/v1/auth/csrf-token`

    ## Response Format
    Success: `{ "success": true, "data": {...} }`
    Error: `{ "success": false, "error": "..." }`

    ## Scouting System
    See `docs/api/scouting-system-api.md` for the scouting system redesign including Prospects,
    20-80 grading scale, present/future grades, and recruiting pipeline.

    ## Stats & Analytics
    See `docs/api/stats-analytics-api.md` for player season/career stats, videos, leaderboards,
    and news releases (all populated by PrestoSports sync).

    ## Extended Stats & Coach's Dashboard
    See `docs/api/extended-stats-api.md` for split stats, raw Presto data, player/team game logs,
    team aggregate stats, lineup, and the coach's dashboard endpoint.
  version: 1.0.0
  contact:
    name: Sports2 Team
servers:
  - url: http://localhost:5000
    description: Local development (reference only; each deployment has its own endpoint)
security:
  - bearerAuth: []
paths:
...
  /api/v1/integrations/presto/league-teams:
    get:
      summary: All league teams with logos
      description: |
        Returns all teams in the configured PrestoSports season with team names and logo URLs.
        Logo URLs point to static.prestosports.com and can be used directly by the frontend.
        Optionally pass `seasonId` query param to override the configured season.
      parameters:
        - name: seasonId
          in: query
          description: Override the configured season ID
          schema: { type: string }
      responses:
        '200':
          description: Array of league teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/LeagueTeam' }
                  count: { type: integer }
                  seasonId: { type: string }
        '400':
          description: No season configured
...
  /api/v1/integrations/presto/configure:
    post:
      summary: Configure Presto integration
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                credentials: { type: object }
      responses:
        '200':
          description: Configuration updated
...
  /api/v1/integrations/presto/settings:
    put:
      summary: Update Presto settings
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Settings updated
...
  /api/v1/integrations/presto/sync/stats:
    post:
      summary: Sync stats from Presto
      tags: [Integrations]
      responses:
        '200':
          description: Sync initiated
        '500':
          description: Sync failed
...
  /api/v1/integrations/presto/sync/all:
    post:
      summary: Full sync from Presto
      tags: [Integrations]
      description: >-
        Runs all sync operations sequentially (roster, schedule, stats, game log,
        team record, season stats, split stats, team aggregates, career stats,
        player details, player photos, historical stats, player videos, press
        releases, opponent stats). Each sub-sync is individually error-handled;
        failures in one don't stop others. Pass `force: true` to bypass TTL/skip
        logic and re-fetch all data.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
                  description: >-
                    When true, bypasses already-synced checks and TTL-based skip
                    logic for stats, career stats, player details, player photos,
                    historical stats, player videos, and opponent stats.
      responses:
        '200':
          description: Full sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      results:
                        type: object
                        properties:
                          roster: { type: object }
                          schedule: { type: object }
                          stats: { type: object }
                          gameLog: { type: object }
                          teamRecord: { type: object }
                          seasonStats: { type: object }
                          splitStats: { type: object }
                          teamAggregateStats: { type: object }
                          careerStats: { type: object }
                          playerDetails: { type: object }
                          playerPhotos: { type: object }
                          historicalStats: { type: object }
                          playerVideos: { type: object }
                          pressReleases: { type: object }
                          opponentStats: { type: object }
                          errors:
                            type: array
                            items: { type: string }
        '500':
          description: Sync failed
  /api/v1/integrations/presto/sync/history:
    get:
      summary: Get sync history audit log
      tags: [Integrations]
      description: >-
        Retrieves paginated sync operation history for the team. Shows when syncs
        were triggered, by whom, duration, and result. Requires authentication.
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 25, maximum: 100 }
        - in: query
          name: sync_type
          schema:
            type: string
            enum: [roster, schedule, stats, team_record, season_stats, career_stats, full, player_details, player_photos, press_releases, historical_stats, historical_season_stats, player_videos, live_stats]
        - in: query
          name: status
          schema:
            type: string
            enum: [started, completed, partial, failed]
      responses:
        '200':
          description: Sync history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        sync_type: { type: string }
                        status: { type: string, enum: [started, completed, partial, failed] }
                        trigger: { type: string, enum: [manual, scheduled] }
                        triggered_by: { type: string, nullable: true }
                        started_at: { type: string, format: date-time }
                        completed_at: { type: string, format: date-time, nullable: true }
                        duration_ms: { type: integer, nullable: true }
                        items_created: { type: integer }
                        items_updated: { type: integer }
                        items_skipped: { type: integer }
                        items_failed: { type: integer }
                        error_message: { type: string, nullable: true }
                        response_summary: { type: object, nullable: true }
                  pagination:
                    type: object
                    properties:
                      page: { type: integer }
                      limit: { type: integer }
                      total: { type: integer }
                      pages: { type: integer }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Required for POST, PUT, PATCH, DELETE
  schemas:
...
    LeagueTeam:
      type: object
      description: A team from the PrestoSports season with logo URL
      properties:
        teamId: { type: string, description: PrestoSports team ID }
        name: { type: string, description: Team display name }
        logo: { type: string, format: uri, description: Logo URL from static.prestosports.com }
        conference: { type: string, nullable: true }
        division: { type: string, nullable: true }
...